Reorganized the tests. Only TreeInfoTest needs access to the members of Pattern and Matcher with package visibility, so it is the only file in the package com.florian.ingerl.util.regex.
The other tests are now in the package
com.florian.ingerl.util.regex.tests. This ensures that these classes
don't have access to the members of Pattern and Matcher with package
visibility.

Added the possibility for the user to install plugins into the regex
engine. This is done by calling Pattern.installPlugin.
This feature still needs to get documented.